app/requirements.txt
gpytranslate==2.0.0
openai==1.57.3
python-dotenv==1.0.1
aiohttp==3.10
redis==4.5.0
aiogram==3.15.0
psycopg-binary==3.2.3
psycopg==3.2.3
psycopg-pool==3.2.4
aiocryptopay==0.4.7
tiktoken==0.8.0
fastapi==0.115.6
uvicorn==0.32.1
gunicorn==20.1.0
pydantic==2.9